[stable5.2] refactor: migrate to `@nextcloud/dialogs` v7

+++ src/components/AppNavigationForm.vue
@@ -98,8 +98,8 @@
 
 <script>
 import { getCurrentUser } from '@nextcloud/auth'
+import { showConfirmation, showError } from '@nextcloud/dialogs'
 import { generateOcsUrl } from '@nextcloud/router'
-import { DialogSeverity, getDialogBuilder, showError } from '@nextcloud/dialogs'
 import NcActionButton from '@nextcloud/vue/components/NcActionButton'
 import NcActionRouter from '@nextcloud/vue/components/NcActionRouter'
 import NcActionSeparator from '@nextcloud/vue/components/NcActionSeparator'
@@ -283,30 +283,18 @@ export default {
 		},
 
 		async onConfirmDelete() {
-			const dialog = getDialogBuilder(t('forms', 'Delete form'))
-				.setText(
-					t('forms', 'Are you sure you want to delete {title}?', {
-						title: this.formTitle,
-					}),
-				)
-				.setSeverity(DialogSeverity.Error)
-				.setButtons([
-					{
-						label: t('forms', 'Cancel'),
-						callback: () => {},
-						variant: 'secondary',
-					},
-					{
-						label: t('forms', 'Delete form'),
-						callback: () => {
-							this.onDeleteForm()
-						},
-						variant: 'error',
-					},
-				])
-				.build()
+			const shouldDelete = await showConfirmation({
+				name: t('forms', 'Delete form'),
+				text: t('forms', 'Are you sure you want to delete {title}?', {
+					title: this.formTitle,
+				}),
+				labelConfirm: t('forms', 'Delete form'),
+				labelReject: t('forms', 'Cancel'),
+			})
 
-			await dialog.show()
+			if (shouldDelete) {
+				await this.onDeleteForm()
+			}
 		},
 
 		async onToggleArchive() {
